pyproject.toml
[tool.poetry]
name = "django-api"
version = "0.1.0"
description = ""
authors = ["eduardo machado <eduardoshy1@gmail.com>"]
readme = "README.md"
packages = [{include = "django_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.0.7"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-admin-interface = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
